+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changes
 * Remove HockeyKit task
 * Removed ReleaseNotes task because the pegdown dependencies causes troubles: Issue #431
 * Added support for skipped tests, that they are not shown as failures.
+* Added carthage archive task, so that a binary framework can be build.
 
 # 0.20.0 (2 June, 2020)